+++ src/app/locales/en/translations.js
@@ -27,7 +27,8 @@ import filesystemBrowser from './components/filesystem-browser';
 import archiveFilesystemBrowser from './components/archive-filesystem-browser';
 import datasetBrowser from './components/dataset-browser';
 import archiveBrowser from './components/archive-browser';
-import archiveSettings from './components/archive-settings';
+import archiveProperties from './components/archive-properties';
+import archiveCreate from './components/archive-create';
 import archivesPurge from './components/archives-purge';
 import itemsSelectBrowser from './components/items-select-browser';
 import datasetProtection from './components/dataset-protection';
@@ -36,7 +37,7 @@ import archiveRecall from './components/archive-recall';
 import fileRecall from './components/file-recall';
 import archiveContentCount from './components/archive-content-count';
 import archiveState from './components/archive-state';
-import archiveCreateModal from './components/archive-create-modal';
+import archivePropertiesModal from './components/archive-properties-modal';
 
 import uploadManager from './services/upload-manager';
 
@@ -50,6 +51,8 @@ import archiveFilesystemBrowserModel from './utils/archive-filesystem-browser-mo
 import itemsSelectBrowserUtils from './utils/items-select-browser';
 import workflowActions from './utils/workflow-actions';
 import workflowVisualiser from './utils/workflow-visualiser';
+import archiveFormBaseModel from './utils/archive-form/-base-model';
+import archiveFormViewModel from './utils/archive-form/view-model';
 
 const translations = {
   components: {
@@ -79,7 +82,8 @@ const translations = {
     archiveFilesystemBrowser,
     datasetBrowser,
     archiveBrowser,
-    archiveSettings,
+    archiveProperties,
+    archiveCreate,
     archivesPurge,
     itemsSelectBrowser,
     datasetProtection,
@@ -88,7 +92,7 @@ const translations = {
     fileRecall,
     archiveContentCount,
     archiveState,
-    archiveCreateModal,
+    archivePropertiesModal,
   },
   services: {
     uploadManager,
@@ -104,6 +108,10 @@ const translations = {
     itemsSelectBrowser: itemsSelectBrowserUtils,
     workflowActions,
     workflowVisualiser,
+    archiveForm: {
+      baseModel: archiveFormBaseModel,
+      viewModel: archiveFormViewModel,
+    },
   },
 };